Rename rest of AdminAuctionStatusPresenter classes
Renamed the Accepted, Rejected and PendingAcceptance presenters to be
within the BidStatusPresenter::Over::Admin namespace
