+++ app/models/admin_auction_status_presenter_factory.rb
@@ -1,3 +1,8 @@
+# this is required because Rails autoloader is a bit confused, will replace soon
+require 'bid_status_presenter/over/admin/rejected'
+require 'bid_status_presenter/over/admin/accepted'
+require 'bid_status_presenter/over/admin/pending_acceptance'
+
 class AdminAuctionStatusPresenterFactory
   attr_reader :auction
 
@@ -13,7 +18,7 @@ class AdminAuctionStatusPresenterFactory
     elsif work_in_progress?
       BidStatusPresenter::Over::Admin::WorkInProgress.new(auction: auction)
     elsif !auction.pending_delivery?
-      Object.const_get("AdminAuctionStatusPresenter::#{status}").new(auction: auction)
+      Object.const_get("BidStatusPresenter::Over::Admin::#{status}").new(auction: auction)
     else
       Object.const_get("C2StatusPresenter::#{c2_status}").new(auction: auction)
     end
