--- app/presenters/admin_auction_status_presenter/rejected.rb
@@ -1,20 +1,0 @@
-class AdminAuctionStatusPresenter::Rejected < AdminAuctionStatusPresenter::Base
-  def header
-    I18n.t('statuses.admin_auction_status_presenter.rejected.header')
-  end
-
-  def body
-    I18n.t(
-      'statuses.admin_auction_status_presenter.rejected.body',
-      delivery_url: auction.delivery_url,
-      rejected_at: rejected_at,
-      winner_email: winner.email
-    )
-  end
-
-  private
-
-  def rejected_at
-    DcTimePresenter.convert_and_format(auction.rejected_at)
-  end
-end
