+++ config/locales/statuses/en.yml
@@ -111,22 +111,22 @@ en:
               %{winner_url} initiated a <a href=%{delivery_url}>pull
               request</a>, and should complete their work no later than
               %{delivery_deadline}.
-    admin_auction_status_presenter:
-      pending_acceptance:
-        body: >
-          %{winner_url} made a <a href=%{delivery_url}>pull request</a> and marked this auction as
-          delivered. If their work meets this auction's acceptance criteria, click
-          "accept." Otherwise, click "reject."
-        header: "Pending acceptance"
-        actions:
-          accept: "Accept"
-          reject: "Reject"
-      accepted:
-        header: "Accepted"
+          pending_acceptance:
+            body: >
+              %{winner_url} made a <a href=%{delivery_url}>pull request</a> and marked this auction as
+              delivered. If their work meets this auction's acceptance criteria, click
+              "accept." Otherwise, click "reject."
+            header: Pending acceptance
+            actions:
+              accept: Accept
+              reject: Reject
+          accepted:
+            header: "Accepted"
         body: "%{winner_email}'s delivered code was accepted on %{accepted_at}."
       rejected:
         header: "Rejected"
         body: "<a href=%{delivery_url}>%{winner_email}'s work</a> was rejected on %{rejected_at}."
+
     c2_presenter:
       approval_not_requested:
         status: "Not requested"
