+++ features/step_definitions/auction_form_steps.rb
@@ -13,11 +13,11 @@ Then(/^I submit a blank delivery URL$/) do
 end
 
 When(/^I mark the auction as rejected$/) do
-  button = I18n.t('statuses.admin_auction_status_presenter.pending_acceptance.actions.reject')
+  button = I18n.t('statuses.bid_status_presenter.over.admin.pending_acceptance.actions.reject')
   step("I click on the \"#{button}\" button")
 end
 
 When(/^I mark the auction as accepted$/) do
-  button = I18n.t('statuses.admin_auction_status_presenter.pending_acceptance.actions.accept')
+  button = I18n.t('statuses.bid_status_presenter.over.admin.pending_acceptance.actions.accept')
   step("I click on the \"#{button}\" button")
 end
